[new] Skin and Skeleton can now be imported [new] Skin and Skeleton can now be set in the AnimationComponent [bug] too few fps when playing animations [todo] Animation should be saved and imported

+++ engine/include/Engine/Serialization/binary/BinarySaver.hpp
@@ -45,7 +45,7 @@ void save(BinarySaver& context, const T& saved, const BinarySaver::SaveInfo* inf
 template <>
 inline void save(BinarySaver& context, const std::string& saved, const BinarySaver::SaveInfo* info)
 {
-    context.saveAsBinary(saved.size() + 1);
+    context.saveAsBinary(size_t(saved.size() + 1));
     context.saveArray(saved.data(), saved.size() + 1);
 }
